+++ src/components/Rules.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { RotateCw } from 'react-feather';
+import { useTranslation } from 'react-i18next';
 import { queryCache, useQuery } from 'react-query';
 import { areEqual, VariableSizeList } from 'react-window';
 import { useRecoilState } from 'recoil';
@@ -8,6 +9,8 @@ import { fetchRules } from 'src/api/rules';
 import { RuleProviderItem } from 'src/components/rules/RuleProviderItem';
 import { TextFilter } from 'src/components/rules/TextFilter';
 import { ruleFilterText } from 'src/store/rules';
+import { State } from 'src/store/types';
+import type { ClashAPIConfig } from 'src/types';
 
 import useRemainingViewPortHeight from '../hooks/useRemainingViewPortHeight';
 import { getClashAPIConfig } from '../store/app';
@@ -21,7 +24,7 @@ const { memo, useMemo, useCallback } = React;
 
 const paddingBottom = 30;
 
-function itemKey(index, { rules, provider }) {
+function itemKey(index: number, { rules, provider }) {
   const providerQty = provider.names.length;
 
   if (index < providerQty) {
@@ -32,7 +35,7 @@ function itemKey(index, { rules, provider }) {
 }
 
 function getItemSizeFactory({ provider }) {
-  return function getItemSize(idx) {
+  return function getItemSize(idx: number) {
     const providerQty = provider.names.length;
     if (idx < providerQty) {
       // provider
@@ -66,16 +69,20 @@ const Row = memo(({ index, style, data }) => {
   );
 }, areEqual);
 
-const mapState = (s) => ({
+const mapState = (s: State) => ({
   apiConfig: getClashAPIConfig(s),
 });
 
 export default connect(mapState)(Rules);
 
-function useRuleAndProvider(apiConfig) {
-  const { data: rules } = useQuery(['/rules', apiConfig], fetchRules, {
-    suspense: true,
-  });
+function useRuleAndProvider(apiConfig: ClashAPIConfig) {
+  const { data: rules, isFetching } = useQuery(
+    ['/rules', apiConfig],
+    fetchRules,
+    {
+      suspense: true,
+    }
+  );
   const { data: provider } = useQuery(
     ['/providers/rules', apiConfig],
     fetchRuleProviders,
@@ -84,11 +91,12 @@ function useRuleAndProvider(apiConfig) {
 
   const [filterText] = useRecoilState(ruleFilterText);
   if (filterText === '') {
-    return { rules, provider };
+    return { rules, provider, isFetching };
   } else {
     const f = filterText.toLowerCase();
     return {
       rules: rules.filter((r) => r.payload.toLowerCase().indexOf(f) >= 0),
+      isFetching,
       provider: {
         byName: provider.byName,
         names: provider.names.filter((t) => t.toLowerCase().indexOf(f) >= 0),
@@ -114,10 +122,12 @@ function Rules({ apiConfig }) {
   // @ts-expect-error ts-migrate(2345) FIXME: Argument of type '{ rules: RuleItem[]; provider: {... Remove this comment to see the full error message
   const getItemSize = getItemSizeFactory({ rules, provider });
 
+  const { t } = useTranslation();
+
   return (
     <div>
       <div className={s.header}>
-        <ContentHeader title="Rules" />
+        <ContentHeader title={t('Rules')} />
         <TextFilter />
       </div>
       {/* @ts-expect-error ts-migrate(2322) FIXME: Type 'number | MutableRefObject<any>' is not assig... Remove this comment to see the full error message */}
@@ -138,7 +148,7 @@ function Rules({ apiConfig }) {
       <Fab
         icon={refreshIcon}
         text="Refresh"
-        position={fabPosition}
+        style={fabPosition}
         onClick={invalidateQueries}
       />
     </div>